agent/pyproject.toml
[project]
name = "ups-guard-agent"
version = "1.0.0"
description = "UPS Guard Agent - 远程关机客户端"
authors = [{name = "KingBoyAndGirl"}]
requires-python = ">=3.11"
dependencies = [
    "websockets>=12.0",
    "pystray>=0.19",
    "Pillow>=10.0",
    "psutil>=5.9",
]

[project.scripts]
ups-guard-agent = "ups_guard_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ups_guard_agent"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pyinstaller>=6.0",
]
